requirements.txt
# asgiref==3.7.2
# certifi==2024.2.2
# cffi==1.16.0
# charset-normalizer==3.3.2
# cryptography==42.0.4
# defusedxml==0.7.1
# dj-database-url==2.1.0
# Django==5.0.2
# django-allauth==0.61.1
# django-heroku==0.3.1
# gunicorn==21.2.0
# idna==3.6
# oauthlib==3.2.2
# packaging==23.2
# psycopg2==2.9.9
# pycparser==2.21
# PyJWT==2.8.0
# python3-openid==3.2.0
# requests==2.31.0
# requests-oauthlib==1.3.1
# sqlparse==0.4.4
# typing_extensions==4.9.0
# tzdata==2024.1
# urllib3==2.2.1
# whitenoise==6.6.0

gunicorn
django-heroku
dj-database-url
django-allauth==0.61.1
boto3
django-storages
django-bootstrap-v5
reportlab
whitenoise
